feat: check version before deploy to ease release
